Extract draggable board layer to a file
