+++ src/Items/Board/types.ts
@@ -1,0 +1,6 @@
+import { ColumnDef, Status } from "../types";
+
+export interface BoardState {
+  columns: Record<Status, ColumnDef>;
+  columnOrder: Status[];
+}
